Changed Template driven to Reactive-Forms.
Edited the code for the reactive forms.Earlier in Register Form, Username and Email already taken was not included into the Validators and allowed user to submit the data(though it showed pop-up).Now validation for already present is included in the Validators (So it wont let user submit untill correct data is passed according to the validations). Now the eror texts are updated rectively on the form instead of going imperative way. Individual test cases for password is being tested now.